% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testable.R
\name{as.testable}
\alias{as.testable}
\title{R ad-hoc testable class}
\usage{
as.testable(x, ...)
}
\arguments{
\item{x}{object to build testable from}

\item{...}{extra arguments}
}
\description{
R ad-hoc testable class
}

