% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{command}
\alias{command}
\title{Build a commmand}
\usage{
command(generator, execute, require = function(state, input) T,
  update = function(state, input, output) state, ensure = function(state,
  input, output) T)
}
\arguments{
\item{execute}{A function from the concrete input,
which executes the true function and returns
concrete output.}

\item{require}{A function from the current (symbolic)
state to a bool, indicating if action is currently
applicable}

\item{update}{A function from state to state, which is
polymorphic over symbolic and concrete inputs and
outputs ( as it is used in both action generation and
command execution ).}

\item{ensure}{A post-condition for a command that must be
verified for the command to be considered a success.}

\item{gen}{A generator which provides random arguments
for a command, given the current (symbolic) state.
If nothing can be done with the current state, one
should preclude the situation with a requires and
return NULL. Otherwise, it should be a list of
arguments (the empty list is ok for functions which
take no arguments).}
}
\description{
Build a commmand
}

