% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{expect_sequential}
\alias{expect_sequential}
\title{Execute a state machine model}
\usage{
expect_sequential(initial.state, actions)
}
\arguments{
\item{initial.state}{the starting state to
build from which is appropriate for this
state machine generator.}

\item{actions}{the list of actions which
are to be run.}
}
\value{
an expectation.
}
\description{
Executes the list of commands sequentially,
ensuring that all postconditions hold.
}
