% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{find.smallest}
\alias{find.smallest}
\title{Search through the trees to find the smallest value we can
which still fails the test.}
\usage{
find.smallest(tree, property, shrink.limit, shrinks, pb = NULL)
}
\description{
Search through the trees to find the smallest value we can
which still fails the test.
}

