% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen-sample}
\alias{gen-sample}
\alias{gen.sample}
\alias{gen.sample.int}
\title{Random Sample Generation}
\usage{
gen.sample(x)

gen.sample.int(n)
}
\arguments{
\item{x}{a list or vector to sample from.}

\item{n}{number which is the maximum integer
sampled from.}
}
\description{
These methods are semantically related to
sample and sample.int.
}
\details{
These generators shrinks to the first
value and 1 respectively.
}
\examples{
gen.sample(1:10)   # a number
gen.sample(c(TRUE,FALSE)) # a boolean
gen.sample.int(10) # a number up to 10

}

