% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{gen.action}
\alias{gen.action}
\title{Generator for actions}
\usage{
gen.action(commands, state, counter)
}
\arguments{
\item{commands}{the list of commands which
we can select choose from. Only commands
appropriate for the state will actually be
selected.}

\item{state}{the current state of the system}

\item{counter}{the output variable this action
can write to.}
}
\value{
a list, acc is the new state and counter
  as well as the action generated.
}
\description{
Generator for actions
}

