% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.beta}
\alias{gen.beta}
\title{Generate a float with a gamma distribution}
\usage{
gen.beta(shape1, shape2, ncp = 0)
}
\arguments{
\item{shape1}{same as shape1 in rbeta}

\item{shape2}{same as shape2 in rbeta}

\item{ncp}{same as ncp in rbeta}
}
\description{
Shrinks towards the median value.
}
