% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.c.of}
\alias{gen.c.of}
\title{Generate a vector of primitive values
from a generator.}
\usage{
gen.c.of(number, generator)
}
\arguments{
\item{number}{length of vector to generate}

\item{generator}{a generator used for vector elements}
}
\description{
Generate a vector of primitive values
from a generator.
}

