% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.c}
\alias{gen.c}
\title{Generate a vector of primitive values
from a generator}
\usage{
gen.c(g, from = 1, to = NULL)
}
\arguments{
\item{g}{a generator used for vector elements}

\item{from}{minimum length of the list of
elements}

\item{to}{maximum length of the list of
elements ( defaults to size if NULL )}
}
\description{
Generate a vector of primitive values
from a generator
}

