% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.date}
\alias{gen.date}
\title{Generate a date between the from and to
dates specified.}
\usage{
gen.date(from = as.Date("1900-01-01"), to = as.Date("3000-01-01"))
}
\arguments{
\item{from}{a \code{Date} value}

\item{to}{a \code{Date} value}
}
\description{
Shrinks towards the \code{from} value.
}
\examples{
gen.date()
gen.date( from = as.Date("1939-09-01"), to = as.Date("1945-09-02"))
}
