% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.ensure}
\alias{gen.ensure}
\title{Generator Conditioning}
\usage{
gen.ensure(requires, generator, discard.limit = 100)
}
\arguments{
\item{requires}{a predicate which must hold before
the generated value is used.}

\item{generator}{a generator used for list elements}

\item{discard.limit}{the maximum number of predicate
failures on a single example before this generator
fails. This should be at most 100, as beyond this
level, R may detect the generator to be an infinite
loop and kill it.}
}
\description{
Require a generator's output to hold some property
}
\details{
This function will fail with stop if the discard
limit is passed.

In general, one should try to avoid using this
function, and instead create appropriate values for
one's tests by construction.
}

