% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.example}
\alias{gen.example}
\title{Sample from a generator.}
\usage{
gen.example(g, size = 5)
}
\arguments{
\item{g}{A generator}

\item{size}{The sized example to view}
}
\description{
Sample from a generator.
}
