% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.gamma}
\alias{gen.gamma}
\title{Generate a float with a gamma distribution}
\usage{
gen.gamma(shape, rate = 1, scale = 1/rate)
}
\arguments{
\item{shape}{same as shape in rgamma}

\item{rate}{same as rate in rgamma}

\item{scale}{same as scale in rgamma}
}
\description{
Shrinks towards the median value.
}
