% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.list.of}
\alias{gen.list.of}
\title{Generate a list of values with fixed length}
\usage{
gen.list.of(number, generator)
}
\arguments{
\item{number}{length of list to generate}

\item{generator}{a generator used for list elements}
}
\description{
Generate a list of values with fixed length
}

