% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.list}
\alias{gen.list}
\title{Generate a list of values, with
length bounded by the size parameter.}
\usage{
gen.list(generator, from = 1, to = NULL)
}
\arguments{
\item{generator}{a generator used for list elements}

\item{from}{minimum length of the list of
elements}

\item{to}{maximum length of the list of
elements ( defaults to size if NULL )}
}
\description{
Generate a list of values, with
length bounded by the size parameter.
}

