% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.no.shrink}
\alias{gen.no.shrink}
\title{Stop a generator from shrinking}
\usage{
gen.no.shrink(g)
}
\arguments{
\item{g}{a generator we wish to remove shrinking
from}
}
\description{
Stop a generator from shrinking
}
