% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen}
\alias{gen}
\title{Constructor for a generator.}
\usage{
gen(x)
}
\description{
We're using R's global random number gen.
}
\details{
A gen is essentially a function which takes
a size parameter and returns a lazy tree.
}

