% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.run}
\alias{gen.run}
\title{Run a generator}
\usage{
gen.run(generator, size)
}
\arguments{
\item{generator}{A generator}

\item{size}{The size parameter passed to the
generation functions}
}
\description{
Samples from a generator or list of generators
producing a (single) lazy rose tree.
}
\details{
This is different to calling generarator$unGen(size)
in that it also works on (nested) lists of generators
and pure values.
}
