% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.shrink}
\alias{gen.shrink}
\title{Helper to create a generator with a
shrink function.}
\usage{
gen.shrink(shrinker, g)
}
\arguments{
\item{shrinker}{a function takes an 'a and
returning a vector of 'a.}

\item{g}{a generator we wish to add shrinking
to}
}
\description{
shrinker takes an 'a and returns a vector of 'a.
}
