% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.sized}
\alias{gen.sized}
\title{Sized generator creation}
\usage{
gen.sized(f)
}
\arguments{
\item{f}{the function, taking a size and
returning a generator}
}
\description{
Helper for making a gen with a size parameter.
Pass a function which takes an int and returns
a gen.
}
\examples{
gen.sized ( function(e) gen.element(1:e) )
}
