% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.unif}
\alias{gen.unif}
\title{Generate a float between the from
and to the values specified.}
\usage{
gen.unif(from, to, shrink.median = T)
}
\arguments{
\item{from}{same as from in runif}

\item{to}{same as to in runif}

\item{shrink.median}{whether to shrink
to the middle of the distribution
instead of the low end.}
}
\description{
Shrinks towards the \code{from} value, or
if \code{shrink.median} is on, the middle.
}
\examples{
gen.unif(0, 1) # a float between 0 and 1
}
