% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glass}
\alias{glass}
\title{Unicode sentence corpus}
\format{A character vector of length 150:}
\source{
\url{http://kermitproject.org/utf8.html}
}
\usage{
glass
}
\description{
A dataset containing the sentence "I can eat glass
and it doesn't hurt me." in quite a few languages.
}
\details{
This corpus can be used to test unicode tools and
functions which require a character vector. One
can use this with \code{\link{gen.element}}.
}
\keyword{datasets}

