% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print.report}
\alias{print.report}
\title{Print information about a generator.}
\usage{
\method{print}{report}(report)
}
\arguments{
\item{g}{A generator}
}
\description{
Print information about a generator.
}

