% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{property}
\alias{property}
\title{Label a property}
\usage{
property(desc, prop)
}
\arguments{
\item{A}{short textual description of the property}

\item{A}{property (specified with forall).}
}
\description{
Give a description to a property
}

