% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{reify}
\alias{reify}
\title{Reify Structures}
\usage{
reify(x, env)
}
\arguments{
\item{x}{A structure potentially holding symbolic
variables, which will be replaced with concrete
values from the environment.}

\item{env}{a map from symbolic to concrete
values.}
}
\value{
a reified structure holding concrete
  values.
}
\description{
Convert a symbolic structure to a concrete one,
using the provided environment.
}

