% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Constructor for a report.}
\usage{
report(tests, shrinks, message, counterexample)
}
\arguments{
\item{tests}{the test on which the failure occurred}

\item{shrinks}{how many shrinks were performed}

\item{message}{the error message to print}

\item{counterexample}{the smallest counterexample to print}
}
\description{
Constructor for a report.
}

