% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{run.prop}
\alias{run.prop}
\title{Run a property (with error handling), and turn it
into a testable.}
\usage{
run.prop(property, arguments)
}
\description{
Run a property (with error handling), and turn it
into a testable.
}

