% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{run.tests}
\alias{run.tests}
\title{Hedgehog property runner}
\usage{
run.tests(path, pattern = "\\\\.[rR]$")
}
\arguments{
\item{path}{directory to search for test files}

\item{pattern}{regex pattern to use when searching
for test files.}
}
\description{
Run all hedgehod property tests within a directory
}

