% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink.halves}
\alias{shrink.halves}
\title{Shrink a number by dividing it into halves.}
\usage{
shrink.halves(x)
}
\arguments{
\item{x}{number to produce halves of}
}
\description{
Shrink a number by dividing it into halves.
}
\examples{
shrink.towards(45)
# 22 11  5  2  1
}
