% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink.list}
\alias{shrink.list}
\title{Shrink a list by edging towards the empty list.}
\usage{
shrink.list(xs)
}
\description{
>>> list [1,2,3]
[[],[2,3],[1,3],[1,2]]
}
\details{
>>> list "abcd"
["","cd","ab","bcd","acd","abd","abc"]

/Note we always try the empty list first, as that is the optimal shrink./
list :: [a] -> [[a]]
list xs =
 concatMap
   (\k -> removes k xs)
   (halves $ length xs)
}

