% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink.removes}
\alias{shrink.removes}
\title{Produce permutations of removing num elements from a list.}
\usage{
shrink.removes(num, xs)
}
\arguments{
\item{num}{the number of values to drop}

\item{xs}{the list to shrink}
}
\description{
Produce permutations of removing num elements from a list.
}
