% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink.towards}
\alias{shrink.towards}
\title{Shrink an integral number by edging towards a destination.}
\usage{
shrink.towards(destination)
}
\arguments{
\item{destination}{the value we want to shrink towards.}
}
\description{
Note we always try the destination first, as that is the optimal shrink.
}
\examples{
shrink.towards (0) (100)
# [0,50,75,88,94,97,99]

shrink.towards(500)(1000)
# [500,750,875,938,969,985,993,997,999]

shrink.towards (-50) (-26)
# [-50,-38,-32,-29,-27]
}
