% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{symbolic}
\alias{symbolic}
\title{A symbolic value.
Really, this is just an integer, which we use
as a name for a value which will exist later
in the computation}
\usage{
symbolic(x)
}
\description{
A symbolic value.
Really, this is just an integer, which we use
as a name for a value which will exist later
in the computation
}

