% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testable.R
\name{testable}
\alias{is.testable}
\alias{testable}
\title{Constructor function for a testable value.}
\usage{
testable(type, message = NULL, srcref = NULL)

is.testable(x)
}
\arguments{
\item{type}{the class of the testable}

\item{message}{an optional message}

\item{srcref}{a source code reference (unused)}

\item{x}{object to test for class membership}
}
\description{
These are pretty much just a bool with
a few values for exceptions and warnings.
}
\details{
Partially derived from testthat (MIT licenced).
}

