% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{towards}
\alias{towards}
\title{Shrink an integral number by edging towards a destination.}
\usage{
towards(destination)
}
\description{
>>> towards (0) (100)
 [0,50,75,88,94,97,99]
}
\details{
>>> towards(500)(1000)
 [500,750,875,938,969,985,993,997,999]

>>> towards (-50) (-26)
 [-50,-38,-32,-29,-27]

/Note we always try the destination first, as that is the optimal shrink./
}

