% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree.replicate}
\alias{tree.replicate}
\alias{tree.replicateS}
\title{Creating trees of lists}
\usage{
tree.replicate(num, ma)

tree.replicateS(num, ma, s, ...)
}
\arguments{
\item{num}{the length of the list in the tree}

\item{ma}{a function which (randomly) creates
new tree to add to the list}

\item{s}{a state used when replicating to
keep track of.}

\item{...}{extra arguments to pass to the tree
generating function}
}
\description{
Build a tree of a list, potentially
keeping hold of an internal state.
}
