% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree.replicateM}
\alias{tree.replicateM}
\title{Create a tree of a vector from a generator for a value.}
\usage{
tree.replicateM(num, ma, ...)
}
\description{
Create a tree of a vector from a generator for a value.
}

