% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree.replicateS}
\alias{tree.replicateS}
\title{Create a tree of a vector from a generator for a value.}
\usage{
tree.replicateS(num, ma, s, ...)
}
\description{
Create a tree of a vector from a generator for a value.
}

