% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree.unfoldForest}
\alias{tree.unfoldForest}
\title{Generate a set of tree while there are values in the shrink.}
\usage{
tree.unfoldForest(shrink, x)
}
\description{
Generate a set of tree while there are values in the shrink.
}

