% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{unfoldgenerator}
\alias{unfoldgenerator}
\title{Turn a generator into a tree and a list of generators
into a list of trees.
Non-generator and list values are passed along
as is.
Generators can use the random number generator when
creating their trees.}
\usage{
unfoldgenerator(generator, size)
}
\description{
Turn a generator into a tree and a list of generators
into a list of trees.
Non-generator and list values are passed along
as is.
Generators can use the random number generator when
creating their trees.
}

