% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{vec}
\alias{vec}
\title{Generate a vector of primitive values
from a generator}
\usage{
vec(g)
}
\arguments{
\item{g}{a generator used for vector elements}

\item{number}{length of vector to generate}
}
\description{
Generate a vector of primitive values
from a generator
}

