% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{executeSequential}
\alias{executeSequential}
\title{Execute a set of commands sequentially, ensuring
that all postconditions hold.}
\usage{
executeSequential(initial.state, actions)
}
\description{
Execute a set of commands sequentially, ensuring
that all postconditions hold.
}

