% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{find.smallest}
\alias{find.smallest}
\title{Search through the trees to find the smallest value we can
which still fails the test.}
\usage{
find.smallest(tree, property, shrink.limit, shrinks)
}
\arguments{
\item{tree}{the tree to search through for the smallest
value which fails the test.}

\item{property}{the property which is failing}

\item{shrink.limit}{the limit to how far we will try and shrink}

\item{shrinks}{the current number of shrinks}
}
\description{
Search through the trees to find the smallest value we can
which still fails the test.
}

