% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.c}
\alias{gen.c}
\title{Generate a vector of values from a generator}
\usage{
gen.c(generator, from = 1, to = NULL, of = NULL)
}
\arguments{
\item{generator}{a generator used for vector elements}

\item{from}{minimum length of the list of
elements}

\item{to}{maximum length of the list of
elements (defaults to size if NULL)}

\item{of}{the exact length of the list of
elements (exclusive to `from` and `to`).}
}
\description{
Generate a vector of values from a generator
}
