% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen.list}
\alias{gen.list}
\title{Generate a list of values, with
length bounded by the size parameter.}
\usage{
gen.list(generator)
}
\arguments{
\item{g}{a generator used for list elements}
}
\description{
Generate a list of values, with
length bounded by the size parameter.
}

