% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Constructor for a report.}
\usage{
report(tests, shrinks, message, counterexample)
}
\arguments{
\item{g}{A function from size to tree}
}
\description{
Constructor for a report.
}

