% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{run.prop}
\alias{run.prop}
\title{Run a property (with error handling), and turn it
into a testable.}
\usage{
run.prop(property, arguments, single.argument)
}
\arguments{
\item{property}{the property to test}

\item{arguments}{the generated arguments to the property.}
}
\description{
Run a property (with error handling), and turn it
into a testable.
}

