% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink.list}
\alias{shrink.list}
\title{Shrink a list by edging towards the empty list.}
\usage{
shrink.list(xs)
}
\arguments{
\item{xs}{the list to shrink}
}
\description{
Shrink a list by edging towards the empty list.
}
