% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{symbolic}
\alias{symbolic}
\title{A symbolic value.}
\usage{
symbolic(var)
}
\arguments{
\item{var}{the integer output indicator.}
}
\description{
These values are the outputs of a computation
during the calculations' construction, and
allow a value to use the results of a previous
function.
}
\details{
Really, this is just an integer, which we use
as a name for a value which will exist later
in the computation.
}
